+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For a full diff see [`0.18.0...main`][0.18.0...main].
 
 - Show books in reading streak tooltip
 - Add reading progress bar to library view ([#46](https://github.com/weph/obsidian-bookshelf/issues/46))
+- Add option to group books by tag ([#39](https://github.com/weph/obsidian-bookshelf/issues/39))
 
 ## [`0.18.0`][0.18.0] - 2025-08-20